+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changes
 - launchpad #1205063: no more error if no addons are specified (might
   lead to problems with the default config, though, because of server
   wide modules, but that's supposed to be overridable)
+- launchpad #1213469: openerp scripts: support for entry point arguments
 
 1.7.0 (2013-07-26)
 ------------------